api/requirements.txt
FastAPI
firebase-admin
pydantic
uvicorn
python-jose
passlib[bcrypt]
pydantic[email]
python-dotenv
PyJWT
bcrypt<4.1.0
starlette